feat(preference): show the theme/engineer page in the preference page
    The engineer page can clean-up all the cache used in system
