+++ lib/core.dart
@@ -4,8 +4,10 @@ export 'package:go_router/go_router.dart';
 export 'cores/http.dart';
 export 'cores/misc.dart';
 export 'cores/routes.dart';
-export 'cores/screens.dart';
 export 'cores/storage.dart';
+export 'cores/screens/animations.dart';
+export 'cores/screens/misc.dart';
+export 'cores/screens/view.dart';
 export 'l10n/app_localizations.dart';
 
 // vim: set ts=2 sw=2 sts=2 et:
