+++ lib/cores/screens/view.dart
@@ -1,263 +1,7 @@
-// The miscellaneous widget library of the app.
-import 'dart:math';
+// The view widget library of the app.
 import 'package:flutter/material.dart';
 
-// The placeholder for the app's Work-In-Progress screen
-class WIP extends StatelessWidget {
-  final String? title;
-
-  const WIP({
-    super.key,
-    this.title,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final TextStyle? titleStyle = Theme.of(context).textTheme.headlineLarge;
-
-    return Scaffold(
-      body: SafeArea(
-        child: Center(
-          child: Text(title ?? "Work in Progress", style: titleStyle),
-        ),
-      ),
-    );
-  }
-}
-
-// The flipping card widget that shows a front and back side.
-class Flipping extends StatefulWidget {
-  final Widget child;
-  final Duration duration;
-
-  const Flipping({
-    super.key,
-    required this.child,
-    this.duration = const Duration(milliseconds: 800),
-  });
-
-  @override
-  State<Flipping> createState() => _FlippingState();
-}
-
-class _FlippingState extends State<Flipping> with SingleTickerProviderStateMixin {
-  late final AnimationController controller;
-  late final Animation<double> animation;
-
-  bool flipX = true;
-
-  @override
-  void initState() {
-    super.initState();
-    controller = AnimationController(vsync: this, duration: widget.duration)
-      ..repeat(reverse: true);
-
-    animation = Tween<double>(begin: 0, end: 2 * pi).animate(CurvedAnimation(
-      parent: controller,
-      curve: Curves.easeInOut,
-    ))
-      ..addListener(() => setState(() {}))
-      ..addStatusListener((status) {
-        if (status == AnimationStatus.completed) {
-          flipX = !flipX;
-          controller.reset();
-          controller.forward();
-        }
-      });
-
-    controller.forward();
-  }
-
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final Matrix4 transform = Matrix4.identity();
-    transform.setEntry(3, 2, 0.001);
-    flipX ? transform.rotateY(animation.value) : transform.rotateX(animation.value);
-
-    return AnimatedBuilder(
-      animation: controller,
-      builder: (context, child) {
-        return Transform(
-          transform: transform,
-          alignment: Alignment.center,
-          child: child,
-        );
-      },
-      child: widget.child,
-    );
-  }
-}
-
-// The InkWell wrapper that is no any animation and color.
-class InkWellDone extends StatelessWidget {
-  final Widget child;
-  final VoidCallback? onTap;
-  final VoidCallback? onDoubleTap;
-  final VoidCallback? onLongPress;
-
-  const InkWellDone({
-    super.key,
-    required this.child,
-    this.onTap,
-    this.onDoubleTap,
-    this.onLongPress,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: onTap,
-      onDoubleTap: onDoubleTap,
-      onLongPress: onLongPress,
-      splashColor: Colors.transparent,
-      highlightColor: Colors.transparent,
-      hoverColor: Colors.transparent,
-      child: child,
-    );
-  }
-}
-
-// The clock-like progress indicator which shows the progress of the task.
-class ClockProgressIndicator extends StatefulWidget {
-  final double size;
-  final double barHeight;
-  final double barWidth;
-  final Duration duration;
-  final Color? color;
-
-  const ClockProgressIndicator({
-    super.key,
-    this.size = 40.0,
-    this.barHeight = 10.0,
-    this.barWidth = 3.75,
-    this.duration = const Duration(milliseconds: 650),
-    this.color,
-  });
-
-  @override
-  State<ClockProgressIndicator> createState() => _ClockProgressIndicatorState();
-}
-
-class _ClockProgressIndicatorState extends State<ClockProgressIndicator> with SingleTickerProviderStateMixin {
-  late final AnimationController controller;
-  late final Animation<double> animation;
-
-  @override
-  void initState() {
-    super.initState();
-    controller = AnimationController(
-      vsync: this,
-      duration: widget.duration,
-    )..repeat();
-
-    animation = Tween<double>(begin: 0, end: 1).animate(controller);
-  }
-
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.only(top: 18.0),
-        child: buildContent(),
-      ),
-    );
-  }
-
-  Widget buildContent() {
-    return SizedBox(
-      width: widget.size * 2,
-      height: widget.size * 2,
-      child: AnimatedBuilder(
-        animation: animation,
-        builder: (context, child) => buildClockBar(),
-      ),
-    );
-  }
-
-  Widget buildClockBar() {
-    final Color color = widget.color ?? Theme.of(context).colorScheme.secondary;
-
-    return Stack(
-      children: List.generate(12, (index) {
-        final double angle = (2 * pi / 12) * index;
-        final double radius = widget.size / 2;
-
-        final double x = radius + radius * 0.8 * cos(angle);
-        final double y = radius + radius * 0.8 * sin(angle);
-        final double progress = ((animation.value * 12) - index) % 12 / 12;
-        final Color barColor = color.withValues(alpha: progress);
-
-        return Positioned(
-          left: x - 2,
-          top: y - 2,
-          child: Transform.rotate(
-            angle: angle + pi / 2,
-            child: Container(
-              width: widget.barWidth,
-              height: widget.barHeight,
-              decoration: BoxDecoration(
-                color: barColor,
-                borderRadius: BorderRadius.circular(widget.barWidth / 2),
-              ),
-            ),
-          ),
-        );
-      }),
-    );
-  }
-}
-
-// The widget to show the search result is empty, and show the
-// message to the user.
-class NoResult extends StatelessWidget {
-  final String message;
-  final double size;
-  final IconData icon;
-
-  const NoResult({
-    super.key,
-    this.message = "",
-    this.size = 64,
-    this.icon = Icons.sentiment_dissatisfied_outlined,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: buildContent(context),
-      ),
-    );
-  }
-
-  Widget buildContent(BuildContext context) {
-    final Color color = Theme.of(context).colorScheme.secondary;
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        Icon(icon, size: size, color: color),
-        const SizedBox(height: 8),
-        Text(message, style: Theme.of(context).textTheme.bodyLarge),
-      ],
-    );
-  }
-}
+import 'package:glacial/cores/screens/misc.dart';
 
 // The customized tab view that can be used to show the active and inactive
 // tabs and slide the content to trigger the animation.
